Move cmpLevel function to base namespace
